runpod-worker/requirements.generation.txt
diffusers>=0.32.0,<1.0
transformers>=4.45.0,<5.0
accelerate>=0.34.0,<2.0
safetensors>=0.4.5,<1.0
sentencepiece>=0.2.0,<1.0
Pillow>=10.4.0,<12.0
imageio>=2.36.0,<3.0
